modules/frida_hook.py
import sys
from colorama import Fore, Style

class FridaMemoryHooker:
    """
    Module phân tích động (Dynamic Analysis) dùng Frida.
    Chọc thẳng vào RAM để lấy String (Vector 1.1) và Bypass Anti-Theft (Vector 8.2).
    """
    def __init__(self, target_path):
        self.target_path = target_path
        self.findings = []
        
    def scan(self):
        print(f"{Fore.CYAN}  [-] Khởi tạo Môi trường Dynamic (Frida Instrumentation)...{Style.RESET_ALL}")
        try:
            import frida
            print(f"  {Fore.YELLOW}[*] Cảnh báo: Việc tiêm Frida sẽ mở ứng dụng và có thể bị Defender chặn.{Style.RESET_ALL}")
            # Pseudo-code logic for the actual frida implementation:
            # pid = frida.spawn([self.target_path])
            # session = frida.attach(pid)
            # script = session.create_script("...")
            # script.load()
            # frida.resume(pid)
            
            self.findings.append({
                "id": "DYN-MEM-001",
                "name": "Frida Instrumentation Pipeline Ready",
                "severity": "INFO", 
                "details": "Môi trường Frida đã sẵn sàng. Cần định nghĩa `exports` hooking cụ thể để dump RAM Memory real-time."
            })
            
            print(f"  {Fore.GREEN}[OK] Frida Module đã được cấu hình. (Mock Execution){Style.RESET_ALL}")
        except ImportError:
            print(f"  {Fore.RED}[!] Thiếu thư viện frida. Hãy chạy `pip install frida-tools`.{Style.RESET_ALL}")
        except Exception as e:
            print(f"  {Fore.RED}[!] Lỗi Frida Hooking: {str(e)}{Style.RESET_ALL}")
            
        return self.findings
