requirements.txt
pefile>=2023.2.7
yara-python>=4.3.0
frida-tools>=12.1.2
mitmproxy>=9.0.0
psutil>=5.9.5
colorama>=0.4.6
requests>=2.31.0
Jinja2>=3.1.2
pyzipper==0.3.6
pycryptodomex>=3.19.0
